Server/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
supabase==1.2.0
python-dotenv==1.0.0
Werkzeug==2.3.7
PyJWT==2.8.0
